requirements.txt
langchain
langchain-openai
langchain-community
langchain-huggingface
langgraph
chromadb
sentence-transformers
python-telegram-bot
python-dotenv